crawler/requirements.txt
kafka-python==1.4.7
unicorn_binance_websocket_api==2.1.2
pymongo==4.6.0